+++ src/main/java/ru/kata/spring/boot_security/demo/service/UserService.java
@@ -8,8 +8,8 @@ import java.util.List;
 
 public interface UserService extends UserDetailsService {
     public List<User> index();
-    public User show(int id);
+    public User show(Long id);
     public boolean save(User user);
     public void update(User updatedUser);
-    public void delete(int id);
+    public void delete(Long id);
 }
